refactor: renames pages index files
