+++ src/pages/app/consumer/cart/cart.ts
@@ -1,0 +1,5 @@
+import { Address } from './address';
+import { Payment } from './payment';
+import { Products } from './products';
+
+export const Cart = { Address, Payment, Products };
