+++ src/pages/app/consumer/cart/index.ts
@@ -1,5 +1,1 @@
-import { Address } from './address';
-import { Payment } from './payment';
-import { Products } from './products';
-
-export const Cart = { Address, Payment, Products };
+export { Cart } from './cart';
