+++ src/pages/app/consumer/index.ts
@@ -1,10 +1,10 @@
-export * from './cart';
-export * from './market';
-export * from './markets';
-export * from './producer';
-export * from './producers';
-export * from './product';
-export * from './payment-methods';
-export * from './transactions-in-progress';
-export * from './concluded-transactions';
-export * from './transaction';
+export { Cart } from './cart';
+export { Market } from './market';
+export { Markets } from './markets';
+export { Producer } from './producer';
+export { Producers } from './producers';
+export { Product } from './product';
+export { PaymentMethods } from './payment-methods';
+export { TransactionsInProgress } from './transactions-in-progress';
+export { ConcludedTransactions } from './concluded-transactions';
+export { Transaction } from './transaction';
