+++ src/pages/app/producer/index.ts
@@ -1,11 +1,11 @@
-export * from './pending-transactions';
-export * from './product';
-export * from './products';
-export * from './transaction';
-export * from './transactions';
-export * from './transactions-separation';
-export * from './transactions-in-progress';
-export * from './concluded-transactions';
-export * from './stall';
-export * from './management-products';
-export * from './management-markets';
+export { PendingTransactions } from './pending-transactions';
+export { Product } from './product';
+export { Products } from './products';
+export { Transaction } from './transaction';
+export { Transactions } from './transactions';
+export { TransactionsSeparation } from './transactions-separation';
+export { TransactionsInProgress } from './transactions-in-progress';
+export { ConcludedTransactions } from './concluded-transactions';
+export { Stall } from './stall';
+export { ManagementProducts } from './management-products';
+export { ManagementMarkets } from './management-markets';
